pyproject.toml
[project]
name = "candidate-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "altair>=6",
    "anthropic>=0.77.0",
    "boto3>=1.42.40",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "openai>=2.16.0",
    "pandas<3.0.0",
    "pgvector>=0.4.2",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "streamlit>=1.54.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]

[tool.setup]
packages = ["Chatbot", "database"]
